machine_learning/rag_system/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
langchain-text-splitters>=0.0.1

google-generativeai>=0.3.0
supabase>=2.0.0
vecs>=0.4.0

PyPDF2>=3.0.1
python-docx>=0.8.11
python-dotenv>=1.0.0
unstructured>=0.10.0
